code shifting
moved pull & pall to own file\n call prototype update, stack is not double pointer
