+++ monty.h
@@ -39,7 +39,7 @@ typedef struct instruction_s
 } instruction_t;
 
 stack_t *add_node_start(stack_t **h, int n);
-int call(char **tokens, stack_t *stack);
+int call(char **tokens, stack_t **stack);
 void push(stack_t **, unsigned int n);
 void pall(stack_t **, unsigned int n);
