+++ op_call.c
@@ -1,59 +1,12 @@
 #include "monty.h"
 
-/**
- * push - pushes a node to a stack
- * @stack: head of the stack
- * @n: line number
- */
-void push(stack_t **stack, unsigned int n)
-{
-	stack_t *node = NULL;
-	(void) n;
-
-	node = malloc(sizeof(stack_t));
-	if (!node)
-	{
-		fprintf(stderr, "Error: malloc failed\n");
-		exit(EXIT_FAILURE);
-	}
-
-	node->prev = node->next = NULL;
-
-	if (!(*stack))
-		(*stack) = node;
-	else
-	{
-		(*stack)->prev = node;
-		node->next = *stack;
-		*stack = node;
-	}
-}
 
 /**
- * pall - prints a stack
- * @stack: pointer to a stack_t
- * @n: line number
- */
-void pall(stack_t **stack, unsigned int n)
-{
-	(void) n;
-
-	if (!stack || !(*stack))
-		return;
-
-	while (*stack)
-	{
-		printf("%d\n", (*stack)->n);
-		(*stack) = (*stack)->next;
-  }
-}
-
-
  * call - call appropriate function
  * @t: pointer to array of tokens
  * Return: nothing
  */
-void call(char **tokens, stack_t *stack)
+void call(char **tokens, stack_t **stack)
 {
 	int idx = 0;
 	int line_number = atoi(*t[1]);
@@ -75,6 +28,5 @@ void call(char **tokens, stack_t *stack)
 	}
 
 	if (strcmp(ops[idx].opcode, "push") == 0)
-	{
-
+		(*stack)->n = atoi(*token[1])
 }
